FollowListViewModelをインタフェースと実装に分ける
